+++ frontend/src/pages/Analysis.tsx
@@ -2,14 +2,9 @@ import { NavLink, useNavigate } from "react-router-dom";
 import { useRef, useCallback } from "react";
 import VideoPlayer from "../components/VideoPlayer";
 import type { CloudinaryPlayer } from "../utils/cloudinaryService";
-import {
-  addRehearsal,
-  getCurrentRehearsal,
-  getCurrentSpeech,
-  deleteCurrentRehearsal,
-  deleteSpeech,
-} from "../utils/auth";
+import { addRehearsal, getCurrentSpeech } from "../utils/auth";
 import axios from "axios";
+import ExitButton from "../components/ExitButton";
 
 function Analysis() {
   const playerRef = useRef<CloudinaryPlayer | null>(null);
@@ -45,37 +40,9 @@ function Analysis() {
     }
   };
 
-  const handleExit = async () => {
-    try {
-      const rehearsalId = getCurrentRehearsal();
-      const speechId = getCurrentSpeech();
-
-      if (!rehearsalId || !speechId) {
-        throw new Error("No rehearsal or speech ID found");
-      }
-
-      const rehearsalDeleted = await axios.delete(
-        `http://localhost:8000/rehearsal/${rehearsalId}`,
-      );
-      const speechDeleted = await axios.delete(
-        `http://localhost:8000/speech/${speechId}`,
-      );
-
-      if (rehearsalDeleted.status === 200 && speechDeleted.status === 200) {
-        deleteCurrentRehearsal();
-        deleteSpeech(speechId);
-        navigate("/");
-      }
-    } catch (error) {
-      console.error("Error deleting speech:", error);
-    }
-  };
-
   return (
     <div className="layout-tb">
-      <button onClick={handleExit} className="justify-self-end">
-        exit
-      </button>
+      <ExitButton />
       <div className="grid h-full grid-cols-2 gap-4">
         <div>
           {/* Jump to Middle button is here, outside of VideoPlayer */}
