+++ frontend/src/pages/Content.tsx
@@ -1,12 +1,7 @@
 import { NavLink, useNavigate, useParams } from "react-router-dom";
 import { useState } from "react";
 import axios from "axios";
-import {
-  deleteSpeech,
-  deleteCurrentRehearsal,
-  getCurrentRehearsal,
-  getCurrentSpeech,
-} from "../utils/auth";
+import ExitButton from "../components/ExitButton";
 
 function Content() {
   const [content, setContent] = useState("");
@@ -38,37 +33,9 @@ function Content() {
     }
   };
 
-  const handleExit = async () => {
-    try {
-      const rehearsalId = getCurrentRehearsal();
-      const speechId = getCurrentSpeech();
-
-      if (!rehearsalId || !speechId) {
-        throw new Error("No rehearsal or speech ID found");
-      }
-
-      const rehearsalDeleted = await axios.delete(
-        `http://localhost:8000/rehearsal/${rehearsalId}`,
-      );
-      const speechDeleted = await axios.delete(
-        `http://localhost:8000/speech/${speechId}`,
-      );
-
-      if (rehearsalDeleted.status === 200 && speechDeleted.status === 200) {
-        deleteCurrentRehearsal();
-        deleteSpeech(speechId);
-        navigate("/");
-      }
-    } catch (error) {
-      console.error("Error deleting speech:", error);
-    }
-  };
-
   return (
     <div className="layout-tb">
-      <button onClick={handleExit} className="justify-self-end">
-        exit
-      </button>
+      <ExitButton />
       <div className="flex h-full w-full max-w-4/5 flex-col gap-8 justify-self-center">
         <div className="flex flex-col items-center gap-5 text-center">
           <h1>what are you uploading?</h1>
